split API public/internal
